[UXE-4468] bug: fix: when theme is not defined by input, use perfers-color-scheme (#1604)
When a user accesses the login screen for the first time in a browser (or in an incognito window) and changes the theme before logging in, our store initially triggers the watch as undefined and shortly after with the correct theme.

Since we weren't checking for undefined values, the function didn't know how to handle the azion-undefined theme, resulting in a themeless page. Now, we're checking for invalid values, and if found, we apply the (prefers-scheme-color) as the default.